+++ doc/CHANGES.rst
@@ -31,7 +31,9 @@ CHANGELOG
   the model_validator choosen. If you have only one targeted Model use the
   relationship ManyToOne
 * Reverted the deletion of the core.Session
-* removed dependencies of pkg_ressource for entrypoints and use importlib.
+* With python > 3.11, removed dependencies of pkg_ressource for entrypoints 
+  and use importlib.
+* Added the capability to add mixin to build the Registry Class.
 
 2.1.0 (2023-07-26)
 ------------------